fix bug routes user/verified
