+++ routes/user.routes.js
@@ -1,6 +1,6 @@
 import express from "express";
 import { UserModel } from "../model/user.model.js";
-import { UserVerificationModel } from "../model/userVerification.model.js";
+import { UserVerificationModel } from "../model/userVerification.model.js"; // ->?
 import bcrypt from "bcrypt";
 import * as dotenv from "dotenv";
 import { generateToken } from "../config/jwt.config.js";
@@ -11,11 +11,13 @@ import isTeacher from "../middlewares/isTeacher.js";
 import jwt from "jsonwebtoken";
 import transporter from "../config/transporter.config.js";
 import { v4 as uuidv4 } from "uuid";
-import * as path from "path"
+import * as path from "path";
 
 dotenv.config();
 
 const userRouter = express.Router();
+
+// salt bcrypt
 
 // CREATE USER
 userRouter.post("/signup", async (req, res) => {
@@ -24,24 +26,17 @@ userRouter.post("/signup", async (req, res) => {
     // pass inserida tem que ter carateser etc
     if (
       !password ||
-      !password.match(
-        /^(?=.*?[A-Z])(?=.*?[a-z])(?=.*?[0-9])(?=.*?[#?!@$ %^&*-]).{8,}$/gm
+      !/^(?=.*?[A-Z])(?=.*?[a-z])(?=.*?[0-9])(?=.*?[#?!@$ %^&*-]).{8,}$/gm.test(
+        password // -> correct?
       )
     ) {
       return res.status(400).json({
         msg: "Your password must include minimum eight characters, at least one uppercase letter, one lowercase letter, one number and one special character.",
       });
     }
-    // CRYPTO PASSWORD E GUARDO NO DB UMA HASHEDPASSWORD
 
-    // salt bcrypt
     const salt = await bcrypt.genSalt(Number(process.env.SALT_ROUNDS));
-
-    // cryptografia bcrypt
     const hashedPassword = await bcrypt.hash(password, salt);
-
-    // console.log(`SALT = ${salt}`);
-    // console.log(`HASHED PASSWORD = ${hashedPassword}`);
 
     const createdUser = await UserModel.create({
       ...req.body,
@@ -49,12 +44,13 @@ userRouter.post("/signup", async (req, res) => {
       verified: false,
       // role: "USER" -> NO X TRACHER!!!
       // pra criar novos admin, criar uma rota custom que so um admin pode criar pra criar outros admin
-    }).then((result) => {
-      sendVerificationEmail(result, res);
     });
 
     delete createdUser._doc.passwordHash;
 
+    await sendVerificationEmail(createdUser, res);
+
+    // Error in the createdUser (ERR_HTTP_HEADERS_SENT)
     return res.status(201).json(createdUser);
   } catch (err) {
     console.log(err);
@@ -62,145 +58,102 @@ userRouter.post("/signup", async (req, res) => {
   }
 });
 
-// send verification email
+// SEND EMAIL VERIFICATION
+const sendVerificationEmail = async ({ _id, email }, res) => {
+  try {
+    //  Integrate http://localhost:... 2.08 2. video
+    // OLDER: const currentUrl = Number(process.env.PORT);
+    const currentUrl = `http://localhost:${Number(process.env.PORT)}`;
+    const uniqueString = uuidv4() + _id;
 
-const sendVerificationEmail = ({ _id, email }, res) => {
-  const currentUrl = Number(process.env.PORT);
+    // hash uniqueString
+    const salt = await bcrypt.genSalt(Number(process.env.SALT_ROUNDS));
+    const hashedUniqueString = await bcrypt.hash(uniqueString, salt);
 
-  const uniqueString = uuidv4() + _id;
-
-  const mailOptions = {
-    from: `<${process.env.EMAIL_ADDRESS}>`,
-    to: email,
-    subject: "Verify your email",
-    html: `<p>Verify your email adress to complete the signup and login into your account.</p><p>This link expires in 6 hours.</b></p><p>Press <a href=${
-      currentUrl + "user/verify/" + _id + "/" + uniqueString
-    }>here</a> to proceed.</p>`,
-  };
-
-  // hash the uniqueString
-  // const salt = await bcrypt.genSalt(Number(process.env.SALT_ROUNDS));
-
-  // const hashedUniqueString = await bcrypt.hash(uniqueString, salt);
-
-  const saltRounds = 10;
-  bcrypt
-    .hash(uniqueString, saltRounds)
-    .then((hashedUniqueString) => {
-      const newVerification = UserVerification({
-        userId: _id,
-        uniqueString: hashedUniqueString,
-        createdAt: Date.now(),
-        expiresAt: Date.now() + 21600000,
-      });
-
-      newVerification
-        .save()
-        .then(() => {
-          transporter
-            .sendMail(mailOptions)
-            .then(() => {
-              res.json({
-                status: "PENDING",
-                message: "Verification email sent",
-              });
-            })
-            .catch(error);
-          res.json({
-            status: "FAILED",
-            message: "Verification email failed.",
-          });
-        })
-        .catch((error) => {
-          console.log(error);
-          res.json({
-            status: "FAILED",
-            message: "Couldn't save verification email data.",
-          });
-        });
-    })
-    .catch((error) => {
-      console.log(error);
-      res.json({
-        status: "FAILED",
-        message: "An error occurred while hashing email data.",
-      });
+    const newUserVerification = await UserVerificationModel.create({
+      userId: _id,
+      uniqueString: hashedUniqueString,
+      createdAt: Date.now(),
+      expiresAt: Date.now() + 21600000,
     });
+
+    const sendEmail = await transporter.sendMail({
+      // integrate email value
+      // OLDER: from: `<${process.env.EMAIL_ADDRESS}>`,
+      from: process.env.ETHE_USER,
+      to: email,
+      subject: "Verify your email",
+      text: `Verify your email adress to complete the signup and login into your account. This link expires in 6 hours. Press ${
+        currentUrl + "/user/verify/" + _id + "/" + uniqueString
+      } here to proceed.`,
+      html: `<p>Verify your email adress to complete the signup and login into your account.</p><p>This link expires in 6 hours.</b></p>
+      <p>Press <a href="${
+        currentUrl + `/user/verify/` + _id + `/` + uniqueString
+      }">here</a> to proceed.</p>`,
+    });
+
+    return res.status(201).json({ msg: "Verification email sent" });
+  } catch (err) {
+    console.log(err);
+    res.status(500).json({ msg: "Verification email failed." });
+  }
 };
 
-// verify email
-userRouter.get("/verify/:userId/:uniqueString", (req, res) => {
-  let { userId, uniqueString } = req.params;
-  UserVerification.find({ userId })
-    .then((result)=>{
-      if (result.length > 0){
-        // user verification record exist -> we proceed
-        const {expiresAt} = result[0];
-        const hashedUniqueString = result[0].uniqueString
+// VERIFY EMAIL
+userRouter.get("/verify/:userId/:uniqueString", async (req, res) => {
 
-        //checking for expired unique string
-        if (expiresAt < Date.now()){
-          // record has expired so we delete it
-          UserVerification.deleteOne({userId}).then(result => {
-            User.deleteOne({_id: userId}).then(()=>{
-              let message = "Link has expired. Please sign up again";
-    res.redirect(`/user/verified/error=true&message=${message}`)
-            }).catch((err)=>{
-              let message = "Clearing user with expired unique string failed";
-    res.redirect(`/user/verified/error=true&message=${message}`)
-            })
-          }).catch((err)=>{
-            console.log(err)
-            let message = "An error occurred while clearing user verification record";
-    res.redirect(`/user/verified/error=true&message=${message}`)
-          })
-        
-        } else {
-          // valid record exist -> we validate user string
-          // compare hashed unique string
-          bcrypt.compare(uniqueString, hashedUniqueString).then(result => {
-            if (result){
-              // strings match
+  try {
+    let { userId, uniqueString } = req.params;
+    let verificationModel = await UserVerificationModel.findOne({
+      userId: userId
+    });
+    if (verificationModel) {
+      // user verification record exist -> we proceed
+      const { expiresAt, uniqueString: hashedUniqueString } = verificationModel;
 
-              User.updateOne({_id: userId}, {verified: true}).then(()=>{
-                UserVerification.deleteOne({userId}).then(()=> {
-                  res.sendFile(path.join(_dirname, "./../templates/verified.html"))
-                }).catch(err => {
-                  let message = "An error occurred while finalizing verification ";
-    res.redirect(`/user/verified/error=true&message=${message}`)
-                })
-              }).catch(err => {
-                console.log(err);
-                let message = "An error occurred while updating user record to verified ";
-    res.redirect(`/user/verified/error=true&message=${message}`)
-              })
-
-
-            } else {
-              // existing record but incorrect verification details passed.
-              let message = "Invalid verification details passed. Check your inbox.";
-    res.redirect(`/user/verified/error=true&message=${message}`)
-            }
-          }).catch(err => {
-            let message = "An error occurred while comparing unique strings";
-    res.redirect(`/user/verified/error=true&message=${message}`)
-          })
-        }
+      //checking for expired unique string
+      if (expiresAt < Date.now()) {
+        // record has expired so we delete it
+        await UserVerificationModel.deleteOne({ userId: userId });
+        await UserModel.deleteOne({ _id: userId }); // delete the uses and have to sign up again?
+        let message = "Link has expired. Please sign up again";
+        res.redirect(`/verified/error=true&message=${message}`);
+        // TODO: Missing clearing user
       } else {
-        // user verification record does not exist -> error
-        let message = "An error occurred while chegking for existing user verification record";
-    res.redirect(`/user/verified/error=true&message=${message}`)
-      }
+        let uniqueStringMatch = await bcrypt.compare(
+          uniqueString,
+          hashedUniqueString
+        );
 
-    })
-    .catch((err) => console.log(err));
-    let message = "An error occurred while chegking for existing user verification record";
-    res.redirect(`/user/verified/error=true&message=${message}`)
+        if (uniqueStringMatch) {
+          await UserModel.updateOne({ _id: userId }, { verified: true });
+          await UserVerificationModel.deleteOne({ userId: userId });
+          let message =
+            "You have been successfully verified. Log in to enjoy Yoga Home.";
+          res.redirect(`/verified/message=${message}`);
+          // res.sendFile(path.join(__dirname, "./../templates/verified.html"));
+        } else {
+          let message =
+            "Invalid verification details passed. Check your inbox.";
+          res.redirect(`/verified/error=true&message=${message}`);
+        }
+      }
+    } else {
+      let message =
+        "Account record doesn't exist or has been verified already. Please sign up or log in";
+      res.redirect(`/verified/error=true&message=${message}`);
+    }
+  } catch (err) {
+    console.log(err);
+    let message =
+      "An error occurred while checking for existing user verification record";
+    res.redirect(`/verified/error=true&message=${message}`);
+  }
 });
 
-userRouter.get("/verified", (req, res)=>{
-  res.sendFile(path.join(_dirname, "../templates/verified.html"))
-})
+userRouter.get("/verified", (req, res) => {
+  res.sendFile(path.join(__dirname, "./../templates/verified.html"));
+});
 
 // RECEIVE EMAIL USER
 userRouter.post("/forgot-password", async (req, res) => {
@@ -321,7 +274,7 @@ userRouter.post("/login", async (req, res) => {
       return res.status(404).json({ msg: "Invalid password or email." });
     }
 
-    if(!user.verified) {
+    if (!user.verified) {
       return res.status(404).json({ msg: "Email has not been verified yet." });
     }
